Feat(UI): add trips nav, /trips/[id] detail page with paginated transactions
